+++ NEWS.md
@@ -2,6 +2,8 @@
 - Fixed build on FreeBSD/arm (#534).
 - Fixed crash with generic package instantiation (#520).
 - Now compatible with LLVM 15.0.
+- Fixed calculation of longest static prefix with `'RANGE` expression
+  (#542).
 
 ## Version 1.7.1 - 2022-08-31
 - Added missing textio `WRITE [LINE, REAL, STRING]` in VHDL-2008.